// src/enums/team-roles.enum.ts
export enum TeamRoleEnum {
    ScrumMaster = 'Scrum Master',
    ProductOwner = 'Product Owner',
    Developer = 'Developer',
    QATester = 'QA Tester',
    UXUIDesigner = 'UX/UI Designer',
    TechLead = 'Tech Lead',
    BusinessAnalyst = 'Business Analyst',
    Stakeholder = 'Stakeholder',
    SupportEngineer = 'Support Engineer',
    LEADER = 'LEADER',
  }